double to float for performance

+++ lib/Common/PID_v2.cpp
@@ -17,7 +17,7 @@
  *    The parameters specified here are those for for which we can't set up
  *    reliable defaults, so we need to have the user set them.
  ***************************************************************************/
-PID::PID(double Kp, double Ki, double Kd, int POn, int ControllerDirection)
+PID::PID(float Kp, float Ki, float Kd, int POn, int ControllerDirection)
 {
     // myOutput = Output;
     // myInput = Input;
@@ -39,13 +39,13 @@ PID::PID(double Kp, double Ki, double Kd, int POn, int ControllerDirection)
  *    To allow backwards compatability for v1.1, or for people that just want
  *    to use Proportional on Error without explicitly saying so
  ***************************************************************************/
-PID::PID(double Kp, double Ki, double Kd, int ControllerDirection)
+PID::PID(float Kp, float Ki, float Kd, int ControllerDirection)
     :PID::PID(Kp, Ki, Kd, P_ON_E, ControllerDirection)
 {
 
 }
 
-PID::PID(double Kp, double Ki, double Kd)
+PID::PID(float Kp, float Ki, float Kd)
   :PID::PID(Kp, Ki, Kd, P_ON_E, DIRECT) {
 }
 
@@ -59,17 +59,17 @@ PID::PID()
  *   pid Output needs to be computed.  returns true when the output is computed,
  *   false when nothing has been done.
  **********************************************************************************/
-double PID::Compute(double current_value, double setpoint)
+float PID::Compute(float current_value, float setpoint)
 {
-  double myOutput = outMin - 1.0; //default pid output unavailiable
+  float myOutput = outMin - 1.0; //default pid output unavailiable
   unsigned long now = millis();
   unsigned long timeChange = (now - lastTime);
   if (timeChange >= SampleTime)
   {
     /*Compute all the working error variables*/
-    double input = current_value;
-    double error = setpoint - input;
-    double dInput = (input - lastInput);
+    float input = current_value;
+    float error = setpoint - input;
+    float dInput = (input - lastInput);
     outputSum += (ki * error);
 
     /*Add Proportional on Measurement, if P_ON_M is specified*/
@@ -80,7 +80,7 @@ double PID::Compute(double current_value, double setpoint)
     else if (outputSum < outMin) outputSum = outMin;
 
     /*Add Proportional on Error, if P_ON_E is specified*/
-    double output;
+    float output;
     if (pOnE) output = kp * error;
     else output = 0;
 
@@ -107,7 +107,7 @@ double PID::Compute(double current_value, double setpoint)
  * it's called automatically from the constructor, but tunings can also
  * be adjusted on the fly during normal operation
  ******************************************************************************/
-void PID::SetTunings(double Kp, double Ki, double Kd, int POn)
+void PID::SetTunings(float Kp, float Ki, float Kd, int POn)
 {
    if (Kp<0 || Ki<0 || Kd<0) return;
 
@@ -116,7 +116,7 @@ void PID::SetTunings(double Kp, double Ki, double Kd, int POn)
 
    dispKp = Kp; dispKi = Ki; dispKd = Kd;
 
-   double SampleTimeInSec = ((double)SampleTime)/1000;
+   float SampleTimeInSec = ((float)SampleTime)/1000;
    kp = Kp;
    ki = Ki * SampleTimeInSec;
    kd = Kd / SampleTimeInSec;
@@ -134,7 +134,7 @@ void PID::SetTunings(double Kp, double Ki, double Kd, int POn)
 /* SetTunings(...)*************************************************************
  * Set Tunings using the last-rembered POn setting
  ******************************************************************************/
-void PID::SetTunings(double Kp, double Ki, double Kd)
+void PID::SetTunings(float Kp, float Ki, float Kd)
 {
    SetTunings(Kp, Ki, Kd, P_ON_E);
 }
@@ -146,8 +146,8 @@ void PID::SetSampleTime(int NewSampleTime)
 {
    if (NewSampleTime > 0)
    {
-      double ratio  = (double)NewSampleTime
-                      / (double)SampleTime;
+      float ratio  = (float)NewSampleTime
+                      / (float)SampleTime;
       ki *= ratio;
       kd /= ratio;
       SampleTime = (unsigned long)NewSampleTime;
@@ -162,7 +162,7 @@ void PID::SetSampleTime(int NewSampleTime)
  *  want to clamp it from 0-125.  who knows.  at any rate, that can all be done
  *  here.
  **************************************************************************/
-void PID::SetOutputLimits(double Min, double Max)
+void PID::SetOutputLimits(float Min, float Max)
 {
    if(Min >= Max) return;
    outMin = Min;
@@ -227,17 +227,17 @@ void PID::SetControllerDirection(int Direction)
  * functions query the internal state of the PID.  they're here for display
  * purposes.  this are the functions the PID Front-end uses for example
  ******************************************************************************/
-double PID::GetKp(){ return  dispKp; }
-double PID::GetKi(){ return  dispKi;}
-double PID::GetKd(){ return  dispKd;}
+float PID::GetKp(){ return  dispKp; }
+float PID::GetKi(){ return  dispKi;}
+float PID::GetKd(){ return  dispKd;}
 // int PID::GetMode(){ return  inAuto ? AUTOMATIC : MANUAL;}
 int PID::GetDirection(){ return controllerDirection;}
 
-double PID::GetOutMax() {
+float PID::GetOutMax() {
   return outMax;
 }
 
-double PID::GetOutMin() {
+float PID::GetOutMin() {
   return outMin;
 }
